Fix various bugs across the vm.
* Inconsistent results on repeated execution

When the same expression was executed multiple times, the results would
be inconsistent. This was the result of the state not being reset
correctly when exiting the `run_loop` method.

There are two main points of exit from this method:

1. Successful execution -> OP_RETURN handling code.
2. Error -> early return.

While there was state reset code present, it was being sidestepped in
both cases. Ensuring that this code path was hit, ensured that the state
was being reset properly.

* Inconsisten test execution

When running the same test multiple times, the results were flaky. This
was becuase of my fuckup where I was hardcoding the offsets of values in
the constant store in the tests, when those values were
non-deterministic.

Creating a test-only function which returns all the offsets of all
constants present in the chunk ensures that we are working with the
right values.

* OperatorMismatchError was returning the types in the incorrect order.

This was a function of stack semantics. When type-checking during a
binary operation, we pop the rhs followed by the lhs (as the compilers
encodes it lhs -> rhs in the bytecode), but the error creation code
assumed the opposite was true.